% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggCircle.R
\name{KeggCircle}
\alias{KeggCircle}
\title{A class for representing a circle.}
\description{
A class for representing a circle.

A class for representing a circle.
}
\details{
This class represents a rectangle, used for graphical
representation. It is used by
KeggPathwayConn::extractPathwayMapShapes() method.

Arguments to the constructor are:

x: X coordinate of center.

y: Y coordinate of center.

r: Radius.
}
\examples{
# Create an instance
c1 <- KeggCircle(x=12, y=5, r=3)

# Since it inherits from KeggShape, a color and a label can be set
c2 <- KeggCircle(x=12, y=5, r=3, color='blue', label='Circle 2')

# Getting center
c1$getX()
c1$getY()

# Getting radius
c1#getRadius()

# Draw a circle on the current image
\donttest{
c1$draw()
}

}
\seealso{
\code{\link{KeggShape}}, \code{\link{KeggRect}}.
}
\section{Super class}{
\code{\link[biodbKegg:KeggShape]{biodbKegg::KeggShape}} -> \code{KeggCircle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{KeggCircle$new()}}
\item \href{#method-equals}{\code{KeggCircle$equals()}}
\item \href{#method-getX}{\code{KeggCircle$getX()}}
\item \href{#method-getY}{\code{KeggCircle$getY()}}
\item \href{#method-getRadius}{\code{KeggCircle$getRadius()}}
\item \href{#method-draw}{\code{KeggCircle$draw()}}
\item \href{#method-clone}{\code{KeggCircle$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getColor">}\href{../../biodbKegg/html/KeggShape.html#method-getColor}{\code{biodbKegg::KeggShape$getColor()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getLabel">}\href{../../biodbKegg/html/KeggShape.html#method-getLabel}{\code{biodbKegg::KeggShape$getLabel()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getRgbColor">}\href{../../biodbKegg/html/KeggShape.html#method-getRgbColor}{\code{biodbKegg::KeggShape$getRgbColor()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$new(x, y, r, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-equals"></a>}}
\if{latex}{\out{\hypertarget{method-equals}{}}}
\subsection{Method \code{equals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$equals(other)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getX"></a>}}
\if{latex}{\out{\hypertarget{method-getX}{}}}
\subsection{Method \code{getX()}}{
Get the X coordinate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getX()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The X coordinate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getY"></a>}}
\if{latex}{\out{\hypertarget{method-getY}{}}}
\subsection{Method \code{getY()}}{
Get the Y coordinate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getY()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Y coordinate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-getRadius"></a>}}
\if{latex}{\out{\hypertarget{method-getRadius}{}}}
\subsection{Method \code{getRadius()}}{
Get the radius.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getRadius()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The radius.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-draw"></a>}}
\if{latex}{\out{\hypertarget{method-draw}{}}}
\subsection{Method \code{draw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$draw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
