% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggGenesConn.R
\docType{class}
\name{KeggGenesConn-class}
\alias{KeggGenesConn-class}
\alias{KeggGenesConn}
\title{The connector class to KEGG Pathway database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{BiodbFactory}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getPathwayIds(id, org)}}{:

Gets organism pathways. This method retrieves KEGG pathways of the
    specified organism in which the genes are involved.
    
id: A character vector of KEGG Genes IDs.
    
org: The organism in which to search for pathways, as a KEGG organism code
    (3-4 letters code, like 'hsa', 'mmu', ...). See
    https://www.genome.jp/kegg/catalog/org_list.html for a complete list of KEGG
    organism codes.
    
Returned value: A vector of KEGG pathway IDs.
    }

\item{\code{getPathwayIdsPerGene(id, org, limit = 3)}}{:

Gets organism pathways for each gene. This method retrieves for
    each gene the KEGG pathways of the organism in which the gene is
    involved.
    
id: A character vector of KEGG Gene IDs.
    
org: The organism in which to search for pathways, as a KEGG organism code
    (3-4 letters code, like 'hsa', 'mmu', ...). See
    https://www.genome.jp/kegg/catalog/org_list.html for a complete list of KEGG
    organism codes.
    
limit: The maximum number of modules IDs to retrieve for each gene.
    Set to 0 to disable.
    
Returned value: A named list of KEGG pathway ID vectors, where the names
    of the list are the gene IDs.}
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.genes')

# Get an entry
e <- conn$getEntry('mmu:14635')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{KeggConn}}.
}
