% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggGlycanConn.R
\docType{class}
\name{KeggGlycanConn-class}
\alias{KeggGlycanConn-class}
\alias{KeggGlycanConn}
\title{The connector class to KEGG Glycan database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{BiodbFactory}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}
\section{Methods}{

\describe{
\item{\code{getEntryImageUrl(entry.id)}}{:

Gets the URL to a picture of the entry (e.g.: a picture of the
    molecule in case of a compound entry).
    
entry.id: A character vector containing entry IDs.
    
Returned value: A character vector, the same length as `entry.id`,
    containing for each entry ID either a URL or NA if no URL exists.
    }
}}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Create a connector to KEGG Glycan
conn <- mybiodb$getFactory()$createConn('kegg.glycan')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{KeggConn}}.
}
