% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggGlycanEntry.R
\docType{class}
\name{KeggGlycanEntry-class}
\alias{KeggGlycanEntry-class}
\alias{KeggGlycanEntry}
\title{KEGG Glycan entry class.}
\description{
This is the entry class for the KEGG Glycan database.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.glycan')

# Get an entry
e <- conn$getEntry('G00018')

# Terminate instance.
mybiodb$terminate()

}
