% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggOrthologyConn.R
\docType{class}
\name{KeggOrthologyConn-class}
\alias{KeggOrthologyConn-class}
\alias{KeggOrthologyConn}
\title{The connector class to KEGG Orthology database.}
\description{
This is a concrete connector class. It must never be instantiated directly,
but instead be instantiated through the factory \code{BiodbFactory}.
Only specific methods are described here. See super classes for the
description of inherited methods.
}

\examples{
# Create an instance with default settings:
mybiodb <- biodb::newInst()

# Create a connector
conn <- mybiodb$getFactory()$createConn('kegg.orthology')

# Get an entry
e <- conn$getEntry('K12668')

# Terminate instance.
mybiodb$terminate()

}
\seealso{
\code{\link{KeggConn}}.
}
