% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggRect.R
\docType{class}
\name{KeggRect-class}
\alias{KeggRect-class}
\alias{KeggRect}
\title{A class for representing a rectangle.}
\description{
This class represents a rectangle, used for graphical
representation.
}
\details{
Arguments to the constructor are:

left: Coordinate of left border.

right: Coordinate of right border.

top: Coordinate of top border.

bottom: Coordinate of bottom border.
}
\section{Methods}{

\describe{
\item{\code{draw()}}{:

Draw the shape on the current image.
    
Returned value: None.
    }

\item{\code{equals(other)}}{Test if this shape is the same as another.
    
other: The other shape to compare with.
    
Returned value: TRUE or FALSE.
    }
}}

\examples{
# Create a rectangle instance
r <- KeggRect(left=10, top=10, bottom=20, right=30, color='yellow')

# Draw a rectangle on current image
\donttest{
r$draw()
}

}
\seealso{
\code{\link{KeggShape}}, \code{\link{KeggCircle}}.
}
