% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggCircle.R
\name{KeggCircle}
\alias{KeggCircle}
\title{A class for representing a circle.}
\description{
A class for representing a circle.

A class for representing a circle.
}
\details{
This class represents a rectangle, used for graphical
representation. It is used by
KeggPathwayConn::extractPathwayMapShapes() method.

Arguments to the constructor are:

x: X coordinate of center.

y: Y coordinate of center.

r: Radius.
}
\examples{
# Create an instance
c1 <- KeggCircle$new(x=12, y=5, r=3)

# Since it inherits from KeggShape, a color and a label can be set
c2 <- KeggCircle$new(x=12, y=5, r=3, color='blue', label='Circle 2')

# Getting center
c1$getX()
c1$getY()

# Getting radius
c1#getRadius()

# Draw a circle on the current image
\donttest{
c1$draw()
}

}
\seealso{
\code{\link{KeggShape}}, \code{\link{KeggRect}}.
}
\section{Super class}{
\code{\link[biodbKegg:KeggShape]{biodbKegg::KeggShape}} -> \code{KeggCircle}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KeggCircle-new}{\code{KeggCircle$new()}}
\item \href{#method-KeggCircle-getX}{\code{KeggCircle$getX()}}
\item \href{#method-KeggCircle-getY}{\code{KeggCircle$getY()}}
\item \href{#method-KeggCircle-getRadius}{\code{KeggCircle$getRadius()}}
\item \href{#method-KeggCircle-clone}{\code{KeggCircle$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="draw"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-draw'><code>biodbKegg::KeggShape$draw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="equals"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-equals'><code>biodbKegg::KeggShape$equals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getColor"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getColor'><code>biodbKegg::KeggShape$getColor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getLabel"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getLabel'><code>biodbKegg::KeggShape$getLabel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getRgbColor"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getRgbColor'><code>biodbKegg::KeggShape$getRgbColor()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggCircle-new"></a>}}
\if{latex}{\out{\hypertarget{method-KeggCircle-new}{}}}
\subsection{Method \code{new()}}{
Initialize new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$new(x, y, r, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Abscissa of the circle's center.}

\item{\code{y}}{Ordinate of the circle's center.}

\item{\code{r}}{Radius of the circle.}

\item{\code{...}}{Additional parameters are passed to super class' initializer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggCircle-getX"></a>}}
\if{latex}{\out{\hypertarget{method-KeggCircle-getX}{}}}
\subsection{Method \code{getX()}}{
Get the X coordinate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getX()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The X coordinate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggCircle-getY"></a>}}
\if{latex}{\out{\hypertarget{method-KeggCircle-getY}{}}}
\subsection{Method \code{getY()}}{
Get the Y coordinate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getY()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The Y coordinate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggCircle-getRadius"></a>}}
\if{latex}{\out{\hypertarget{method-KeggCircle-getRadius}{}}}
\subsection{Method \code{getRadius()}}{
Get the radius.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$getRadius()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The radius.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggCircle-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KeggCircle-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggCircle$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
