% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggRect.R
\name{KeggRect}
\alias{KeggRect}
\title{A class for representing a rectangle.}
\description{
A class for representing a rectangle.

A class for representing a rectangle.
}
\details{
This class represents a rectangle, used for graphical
representation.

Arguments to the constructor are:

left: Coordinate of left border.

right: Coordinate of right border.

top: Coordinate of top border.

bottom: Coordinate of bottom border.
}
\examples{
# Create a rectangle instance
r <- KeggRect$new(left=10, top=10, bottom=20, right=30, color='yellow')

# Draw a rectangle on current image
\donttest{
r$draw()
}

}
\seealso{
\code{\link{KeggShape}}, \code{\link{KeggCircle}}.
}
\section{Super class}{
\code{\link[biodbKegg:KeggShape]{biodbKegg::KeggShape}} -> \code{KeggRect}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KeggRect-new}{\code{KeggRect$new()}}
\item \href{#method-KeggRect-clone}{\code{KeggRect$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="draw"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-draw'><code>biodbKegg::KeggShape$draw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="equals"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-equals'><code>biodbKegg::KeggShape$equals()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getColor"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getColor'><code>biodbKegg::KeggShape$getColor()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getLabel"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getLabel'><code>biodbKegg::KeggShape$getLabel()</code></a></span></li>
<li><span class="pkg-link" data-pkg="biodbKegg" data-topic="KeggShape" data-id="getRgbColor"><a href='../../biodbKegg/html/KeggShape.html#method-KeggShape-getRgbColor'><code>biodbKegg::KeggShape$getRgbColor()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggRect-new"></a>}}
\if{latex}{\out{\hypertarget{method-KeggRect-new}{}}}
\subsection{Method \code{new()}}{
Initialize new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggRect$new(left, top, bottom, right, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{left}}{Coordinate of rectangle's left side.}

\item{\code{top}}{Coordinate of rectangle's top side.}

\item{\code{bottom}}{Coordinate of rectangle's bottom side.}

\item{\code{right}}{Coordinate of rectangle's right side.}

\item{\code{...}}{Additional parameters are passed to super class' initializer.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggRect-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KeggRect-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggRect$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
