% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KeggShape.R
\name{KeggShape}
\alias{KeggShape}
\title{A class for representing a shape.}
\description{
A class for representing a shape.

A class for representing a shape.
}
\details{
This abstract class represents a shape, used for graphical representation.

Arguments to the constructor are:

label: A text label to associate with the shape.

color: A color, as a character string.
}
\examples{
# Create a circle instance
c <- KeggCircle$new(x=12, y=5, r=3, label='MyCircle')

# Create a rectangle instance
r <- KeggRect$new(left=10, top=10, bottom=20, right=30, color='yellow')

}
\seealso{
\code{\link{KeggRect}}, \code{\link{KeggCircle}}.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-KeggShape-new}{\code{KeggShape$new()}}
\item \href{#method-KeggShape-equals}{\code{KeggShape$equals()}}
\item \href{#method-KeggShape-getLabel}{\code{KeggShape$getLabel()}}
\item \href{#method-KeggShape-getColor}{\code{KeggShape$getColor()}}
\item \href{#method-KeggShape-getRgbColor}{\code{KeggShape$getRgbColor()}}
\item \href{#method-KeggShape-draw}{\code{KeggShape$draw()}}
\item \href{#method-KeggShape-clone}{\code{KeggShape$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-new"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-new}{}}}
\subsection{Method \code{new()}}{
Initialize new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$new(label = NA_character_, color = NA_character_)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{The text label to display.}

\item{\code{color}}{The color to use.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-equals"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-equals}{}}}
\subsection{Method \code{equals()}}{
Test if this shape is the same as another.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$equals(other)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{other}}{The other shape to compare with.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
TRUE or FALSE.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-getLabel"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-getLabel}{}}}
\subsection{Method \code{getLabel()}}{
Gets the label associated with this shape.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$getLabel()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The label.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-getColor"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-getColor}{}}}
\subsection{Method \code{getColor()}}{
Gets the color associated with this shape.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$getColor()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The color name as a string.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-getRgbColor"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-getRgbColor}{}}}
\subsection{Method \code{getRgbColor()}}{
Gets the RGB color associated with this shape.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$getRgbColor(alpha = 255)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The value to use for the alpha channel when building the RGB
color object.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The color as an RGB color object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-draw"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-draw}{}}}
\subsection{Method \code{draw()}}{
Draw the shape on the current image.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$draw()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
None.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-KeggShape-clone"></a>}}
\if{latex}{\out{\hypertarget{method-KeggShape-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{KeggShape$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
